+++ src/unparse/run.ts
@@ -1,77 +1,81 @@
+import { MapASTNode } from '@superfaceai/ast';
 import fs from 'fs';
+
 import * as unparse from './index';
 
 function nope(): never {
-	console.error('Usage: unparse map|profile --json|--eval --stdin|FILE')
-	process.exit(1)
+  console.error('Usage: unparse map|profile --json|--eval --stdin|FILE');
+  process.exit(1);
 }
 
 function readStdin(): string {
-	const value = fs.readFileSync(0).toString()
+  const value = fs.readFileSync(0).toString();
 
-	if (value.trim() == '') {
-		console.error('Invalid stdin input')
-		nope()
-	}
+  if (value.trim() == '') {
+    console.error('Invalid stdin input');
+    
+	return nope();
+  }
 
-	return value
+  return value;
 }
 function readFile(path: string): string {
-	if (path === undefined || path.trim() === '') {
-		console.error('Invalid file input')
-		nope()
-	}
+  if (path === undefined || path.trim() === '') {
+    console.error('Invalid file input');
+    
+	return nope();
+  }
 
-	return fs.readFileSync(path).toString()
+  return fs.readFileSync(path).toString();
 }
 function readInput(): string {
-	const arg = process.argv[4]
-	switch (arg) {
-		case undefined:
-			nope()
+  const arg = process.argv[4];
+  switch (arg) {
+    case undefined:
+      return nope();
 
-		case '--stdin':
-		case '-':
-			return readStdin()
-		
-		default:
-			return readFile(arg)
-	}
+    case '--stdin':
+    case '-':
+      return readStdin();
+
+    default:
+      return readFile(arg);
+  }
 }
 
 function parseInput(input: string): unknown {
-	const arg = process.argv[3]
-	switch (arg) {
-		case '--eval':
-		case '-e':
-			return eval(input)
-		
-		case '--json':
-		case '-j':
-			return JSON.parse(input)
-		
-		default:
-			nope()
-	}
+  const arg = process.argv[3];
+  switch (arg) {
+    case '--eval':
+    case '-e':
+      return eval(input);
+
+    case '--json':
+    case '-j':
+      return JSON.parse(input);
+
+    default:
+      nope();
+  }
 }
 
 function unparseInput(input: unknown): string {
-	const arg = process.argv[2]
-	switch (arg) {
-		case 'map':
-		case 'm':
-			return new unparse.MapUnparser(input as any).unparse()
-		
-		case 'profile':
-		case 'p':
-			throw 'TODO'
+  const arg = process.argv[2];
+  switch (arg) {
+    case 'map':
+    case 'm':
+      return new unparse.MapUnparser(input as MapASTNode).unparse();
 
-		default:
-			nope()
-	}
+    case 'profile':
+    case 'p':
+      throw 'TODO';
+
+    default:
+      nope();
+  }
 }
 
-const input = readInput()
-const parsed = parseInput(input)
-const unparsed = unparseInput(parsed)
-console.log(unparsed)
+const input = readInput();
+const parsed = parseInput(input);
+const unparsed = unparseInput(parsed);
+console.log(unparsed);